requirements.txt
asgiref==3.4.1
channels==3.0.3
channels-redis==3.3.0
django-eventstream==4.5.1
django-channels-presence==1.0.0
dj-database-url==0.5.0
Django==3.2.17
djangorestframework==3.14.0
django-impersonate==1.7.3
django-mathfilters==1.0.0
django-ratelimit==3.0.1
django-redis==5.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
Markdown==3.3.4
psycopg2==2.9.9
py-cord==2.0.0rc1
requests==2.25.1
Twisted==22.10.0
uvicorn[standard]==0.14.0
whitenoise==5.1.0
python-dotenv==1.0.0
jsonpickle==3.0.2
